Move context_manager.py to token_replacement.py
